hackathon-orchestrator/pyproject.toml
[project]
name = "orch"
version = "0.1.0"
description = "Hackathon agent orchestrator: GitHub Issue -> Cursor CLI -> checks -> repair loop -> PR"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "openai-agents>=0.14.4",
    "openai>=1.54.0",
    "pydantic>=2.9",
    "typer>=0.15",
    "python-dotenv>=1.0.1",
    "httpx>=0.27",
    "rich>=13.9",
]

[project.scripts]
orch = "orch.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["orch*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
# B008 is triggered by Typer's `typer.Option(...)` idiom in function defaults,
# which is the documented way to declare CLI options — so we silence it here.
ignore = ["E501", "B008"]
